Refactor client metrics
