+++ cmd/virt-handler/BUILD.bazel
@@ -10,9 +10,9 @@ go_library(
         "//pkg/container-disk:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/healthz:go_default_library",
-        "//pkg/monitoring/client/prometheus:go_default_library",
         "//pkg/monitoring/domainstats/downwardmetrics:go_default_library",
         "//pkg/monitoring/domainstats/prometheus:go_default_library",
+        "//pkg/monitoring/metrics/virt-handler:go_default_library",
         "//pkg/monitoring/profiler:go_default_library",
         "//pkg/monitoring/reflector/prometheus:go_default_library",
         "//pkg/monitoring/workqueue/prometheus:go_default_library",
