+++ cmd/virt-operator/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
     importpath = "kubevirt.io/kubevirt/cmd/virt-operator",
     visibility = ["//visibility:private"],
     deps = [
-        "//pkg/monitoring/client/prometheus:go_default_library",
         "//pkg/monitoring/reflector/prometheus:go_default_library",
         "//pkg/monitoring/workqueue/prometheus:go_default_library",
         "//pkg/virt-operator:go_default_library",
