+++ pkg/monitoring/metrics/common/client/client_suite_test.go
@@ -1,4 +1,4 @@
-package prometheus
+package client
 
 import (
 	"testing"
