+++ pkg/monitoring/metrics/virt-api/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/monitoring/metrics/virt-api",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/monitoring/metrics/common/client:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//vendor/github.com/machadovilaca/operator-observability/pkg/operatormetrics:go_default_library",
     ],
